+++ crates/compiler/src/builtin/mod.rs
@@ -1,16 +1,25 @@
 mod functions;
 pub(crate) mod modules;
+mod mixins;
 
 pub(crate) use functions::{
     color, list, map, math, meta, selector, string, DISALLOWED_PLAIN_CSS_FUNCTION_NAMES,
     GLOBAL_FUNCTIONS,
 };
 
+#[allow(unused_imports)]
+pub(crate) use mixins::{
+    GLOBAL_MIXINS
+};
+
 pub use functions::Builtin;
+pub use mixins::BuiltinMixin;
 
 /// Imports common to all builtin fns
 mod builtin_imports {
     pub(crate) use super::functions::{Builtin, GlobalFunctionMap, GLOBAL_FUNCTIONS};
+    #[allow(unused_imports)]
+    pub(crate) use super::mixins::{BuiltinMixin, GlobalMixinMap, GLOBAL_MIXINS};
 
     pub(crate) use codemap::{Span, Spanned};