+++ crates/include_sass/src/lib.rs
@@ -6,7 +6,7 @@ use grass_compiler::StdFs;
 use proc_macro::TokenStream;
 #[cfg(not(feature = "nightly"))]
 use quote::format_ident;
-use syn::{parse_macro_input, LitStr};
+use syn::{LitStr, parse_macro_input};
 
 use quote::__private::TokenStream as TokenStream2;